🐛 (fix): Fix forced devices
